Added tests for annotation functions
Using testthat, added tests for annotateNearest(), annnotateTranscripts() and
matchGenes().

Tests for annotateNearest() are based on a manually-created small data set. The
tests check the basic inputs and identifying all scenarios (overlaps, inside,
etc).

Tests for annotateTranscripts() are based on the Canis Familiaris use case from
https://support.bioconductor.org/p/63231/. Basic inputs and some simple expected
output are checked. To use the Cf data, GenomicFeatures is required and hence
added to the suggests field in the DESCRIPTION file.

Tests for matchGenes are also based on the Canis familiaris data. Basic inputs
and scenarios for when strands are changed are checked.